feat: extracted assets and function calls in their own components
